tidy up import of constants
