+++ pallets/exchange/src/lib.rs
@@ -12,9 +12,10 @@ use codec::Encode;
 use sp_std::vec::Vec;
 
 use primitives::{
+	Amount, AssetId, Balance, ExchangeIntention, IntentionType,
 	asset::AssetPair,
 	traits::{Resolver, AMM},
-	Amount, AssetId, Balance, ExchangeIntention, IntentionType, MIN_TRADING_LIMIT,
+	constants::chain::MIN_TRADING_LIMIT,
 };
 use sp_std::borrow::ToOwned;
